cover_drive_analysis_realtime.py
import argparse
import os
import json
import yaml
import cv2

def analyze_video(path: str, output_dir: str, config: dict) -> dict:

    cap = cv2.VideoCapture(path)
    if not cap.isOpened():
        raise FileNotFoundError(f"Could not open video: {path}")


    fps = int(cap.get(cv2.CAP_PROP_FPS))
    width = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
    height = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))


    out_path = os.path.join(output_dir, "annotated_video.mp4")
    fourcc = cv2.VideoWriter_fourcc(*"mp4v")
    writer = cv2.VideoWriter(out_path, fourcc, fps, (width, height))


    frame_count = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
    print(f"Processing {frame_count} frames at {fps} FPS, resolution {width}x{height}")


    while True:
        ret, frame = cap.read()
        if not ret:
            break


        # Dummy pass-through (no overlays yet)
        writer.write(frame)


    cap.release()
    writer.release()


    results = {
    "video": out_path,
    "scores": {},
    "comments": {},
    "metrics_summary": {}
    }
    return results




def main():
    parser = argparse.ArgumentParser(description="AthleteRise Cover Drive Analysis")
    parser.add_argument("--input", type = str, required = True, help = "Path to input video")
    parser.add_argument("--output", type = str, default = "output", help = "Output directory")
    parser.add_argument("--config", type = str, default = "config.yaml", help = "Config file path")
    args = parser.parse_args()


    os.makedirs(args.output, exist_ok = True)


    # Load config
    with open(args.config, "r") as f:
        config = yaml.safe_load(f)


    # Run analysis
    evaluation_dict = analyze_video(args.input, args.output, config)


    # Save evaluation.json
    out_json = os.path.join(args.output, "evaluation.json")
    with open(out_json, "w") as f:
        json.dump(evaluation_dict, f, indent=2)


    print(f"Analysis complete. Results saved in {args.output}")




if __name__ == "__main__":
    main()