requirements.txt
opencv-python
numpy
mediapipe
tqdm
pandas
matplotlib
scipy
pyyaml
pytest